#1 create database design
using draw.io to create an er diagram this should be suitable for the db.

A choice being made here is to assume that images would be saved and managed somewhere else. with that in mind the expectations will be for images to come in the form of imgur links