use request body for filter, sort and page request
I'm choosing to move this from the query string to the request body. I know query data should go in the query string, but to allow for easier parsing and more features with the filter and sort it will be easier to have the request body be used for that.